requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
httpx
pydantic>=2.0
python-multipart
psycopg2-binary
